pyproject.toml
[project]
name = "onchain3r"
version = "0.1.0"
description = "Automated token due diligence system for EVM chains"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "web3>=7.0",
    "anthropic>=0.40",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "typer>=0.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
onchain3r = "onchain3r.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
